Refactored PhenotypeCenterService, splitting out the methods that support the procedureCompleteness report (which require ImpressService and the statisticalResultCore) into a new class, PhenotypeCenterProcedureCompletenessService. This service is used by the reports module and the web module (PhenotypeCenterProgressController.showPhenotypeCenterProgressCsv) and breaking out StrainBean into a new class, PhenotypeCenterServiceBean.
