+++ data-model-solr/src/main/java/org/mousephenotype/cda/solr/service/PhenotypeCenterService.java
@@ -19,7 +19,6 @@ package org.mousephenotype.cda.solr.service;
 import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.apache.solr.client.solrj.SolrServerException;
-import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.apache.solr.client.solrj.response.*;
 import org.apache.solr.client.solrj.response.FacetField.Count;
 import org.apache.solr.common.SolrDocument;
@@ -28,32 +27,24 @@ import org.mousephenotype.cda.solr.service.dto.ObservationDTO;
 import org.mousephenotype.cda.solr.service.dto.ProcedureDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
 
 import java.io.IOException;
-import java.sql.SQLException;
 import java.util.*;
 
+@Service
 public class PhenotypeCenterService {
-	private static final Logger LOG = LoggerFactory.getLogger(PhenotypeCenterService.class);
-	private
-	SolrClient solr;
-	private final String datasourceName = "IMPC";//pipeline but takes care of things like WTSI MGP select is IMPC!
 
-	private final Logger logger = LoggerFactory.getLogger(this.getClass());
+    private final Logger     logger         = LoggerFactory.getLogger(this.getClass());
+    private       SolrClient solrClientExperiment;
+    private final String     datasourceName = "IMPC";//pipeline but takes care of things like WTSI MGP select is IMPC!
 
-	@Autowired
-	private ImpressService impressService;
-
-
-	public PhenotypeCenterService() {
-	}
-
-
-	public PhenotypeCenterService(String baseSolrUrl, ImpressService impressService){
-		solr = new HttpSolrClient(baseSolrUrl);
-		this.impressService = impressService;
-
+	public PhenotypeCenterService(
+			@Qualifier("experimentCore")
+			SolrClient solrClientExperiment)
+	{
+		this.solrClientExperiment = solrClientExperiment;
 	}
 
 	/**
@@ -70,18 +61,18 @@ public class PhenotypeCenterService {
 		.addFacetField(ObservationDTO.PHENOTYPING_CENTER)
 		.setFacetMinCount(1)
 		.setRows(0);
-		if(SolrUtils.getBaseURL(solr).endsWith("experiment")){
+		if(SolrUtils.getBaseURL(solrClientExperiment).endsWith("experiment")){
 			query.addFilterQuery(ObservationDTO.DATASOURCE_NAME+":"+"\""+datasourceName+"\"");
 		}
 
-		QueryResponse response = solr.query(query);
+		QueryResponse response = solrClientExperiment.query(query);
 		//String resp = response.getResponse().toString();
 		List<FacetField> fields = response.getFacetFields();
 		//System.out.println("values="+fields.get(0).getValues());
 		for(Count values: fields.get(0).getValues()){
 			centers.add(values.getName());
 		}
-		//System.out.println("resp="+resp);
+
 		return centers;
 	}
 
@@ -100,21 +91,21 @@ public class PhenotypeCenterService {
 		.setFacetMinCount(1)
 		.setFacetLimit(-1)
 		.setRows(0);
-		if(SolrUtils.getBaseURL(solr).endsWith("experiment")){
+		if(SolrUtils.getBaseURL(solrClientExperiment).endsWith("experiment")){
 				query.addFilterQuery(ObservationDTO.DATASOURCE_NAME + ":" + "\"" + datasourceName + "\"");
 		}
-		QueryResponse response = solr.query(query);
+		QueryResponse response = solrClientExperiment.query(query);
 		List<FacetField> fields = response.getFacetFields();
 		for(Count values: fields.get(0).getValues()){
 			strains.add(values.getName());
 		}
-		logger.info("getStrainsForCenter ---- " + SolrUtils.getBaseURL(solr) + "/select?" + query);
+		logger.info("getStrainsForCenter ---- " + SolrUtils.getBaseURL(solrClientExperiment) + "/select?" + query);
 		return strains;
 	}
 
-	public List<StrainBean> getMutantStrainsForCenter(String center)  throws SolrServerException, IOException  {
+	public List<PhenotypeCenterServiceBean> getMutantStrainsForCenter(String center)  throws SolrServerException, IOException  {
 
-		List<StrainBean> strains=new ArrayList<>();
+		List<PhenotypeCenterServiceBean> strains=new ArrayList<>();
 		SolrQuery query = new SolrQuery()
 			.setQuery(ObservationDTO.PHENOTYPING_CENTER + ":\"" + center + "\" AND " + ObservationDTO.BIOLOGICAL_SAMPLE_GROUP + ":experimental")
 			.setFields(ObservationDTO.GENE_ACCESSION_ID,ObservationDTO.ALLELE_SYMBOL, ObservationDTO.GENE_SYMBOL, ObservationDTO.ZYGOSITY, ObservationDTO.DEVELOPMENTAL_STAGE_NAME)
@@ -124,13 +115,13 @@ public class PhenotypeCenterService {
 		query.set("group.limit", 1);
 
 
-		if(SolrUtils.getBaseURL(solr).endsWith("experiment")){
+		if(SolrUtils.getBaseURL(solrClientExperiment).endsWith("experiment")){
 				query.addFilterQuery(ObservationDTO.DATASOURCE_NAME + ":" + "\"" + datasourceName + "\"");
 		}
-		QueryResponse response = solr.query(query);
+		QueryResponse response = solrClientExperiment.query(query);
 		GroupResponse groups = response.getGroupResponse();
 		for( Group group : groups.getValues().get(0).getValues()){
-			StrainBean strain = new StrainBean();
+			PhenotypeCenterServiceBean strain = new PhenotypeCenterServiceBean();
 			String colonyId = group.getGroupValue();
 			if (colonyId != null && !colonyId.equalsIgnoreCase("null")){
 				strain.setColonyId(colonyId);
@@ -143,7 +134,7 @@ public class PhenotypeCenterService {
 				strains.add(strain);
 			}
 		}
-		LOG.info("getStrainsForCenter -- " + SolrUtils.getBaseURL(solr) + "/select?" + query);
+		logger.info("getStrainsForCenter -- " + SolrUtils.getBaseURL(solrClientExperiment) + "/select?" + query);
 		return strains;
 	}
 
@@ -166,11 +157,11 @@ public class PhenotypeCenterService {
 		 .setFacetMinCount(1)
 		 .setRows(0);
 
-		if(SolrUtils.getBaseURL(solr).endsWith("experiment")){
+		if(SolrUtils.getBaseURL(solrClientExperiment).endsWith("experiment")){
 			query.addFilterQuery(ObservationDTO.DATASOURCE_NAME+":"+"\""+datasourceName+"\"");
 		}
 
-		QueryResponse response = solr.query(query);
+		QueryResponse response = solrClientExperiment.query(query);
 		List<FacetField> fields = response.getFacetFields();
 		for(int i=0; i< fields.get(0).getValues().size(); i++){
 			ProcedureDTO procedure = new ProcedureDTO();
@@ -196,7 +187,7 @@ public class PhenotypeCenterService {
 		 .addFacetPivotField(ObservationDTO.PHENOTYPING_CENTER + "," + ObservationDTO.PROCEDURE_STABLE_ID)
 		 .setFacetMinCount(1)
 		 .setRows(0);
-		QueryResponse response = solr.query(query);
+		QueryResponse response = solrClientExperiment.query(query);
 		Map<String, List<String>> res = new HashMap<>();
 		List<PivotField> fields = response.getFacetPivot().get(ObservationDTO.PHENOTYPING_CENTER + "," + ObservationDTO.PROCEDURE_STABLE_ID);
 
@@ -226,7 +217,7 @@ public class PhenotypeCenterService {
 		 .setFacetMinCount(1)
 		 .setRows(0);
 
-		QueryResponse response = solr.query(query);
+		QueryResponse response = solrClientExperiment.query(query);
 		List<FacetField> fields = response.getFacetFields();
 		for( Count field: fields.get(0).getValues()){
 			procedures.add(field.getName());
@@ -260,72 +251,7 @@ public class PhenotypeCenterService {
 	}
 
 
-	/**
-	 * @author tudose
-	 * @return
-	 * @throws SolrServerException, IOException
-	 * @throws SQLException
-	 */
-	public List<String[]> getCentersProgressByStrainCsv() throws SolrServerException, IOException , SQLException {
-
-		List<String> centers = getPhenotypeCenters();
-        List<String[]> results = new ArrayList<>();
-        String[] temp = new String[1];
-        List<String> header = new ArrayList<>();
-        header.add("Colony Id");
-        header.add("Gene Symbol");
-        header.add("MGI Gene Id");
-        header.add("Allele Symbol");
-
-
-		header.add("Zygosity");
-		header.add("Life Stage");
-        header.add("Phenotyping Center");
-        header.add("Percentage Done");
-        header.add("Number of Done Procedures");
-        header.add("Done Procedures");
-        header.add("MPS Tested For");
-        header.add("Number of Missing Procedures");
-        header.add("Missing Procedures");
-        header.add("MPS Not Tested For");
-		results.add(header.toArray(temp));
-		Map<String, Set<String>> mpsPerProcedure = impressService.getMpsForProcedures();
-		Map<String, List<String>> possibleProceduresPerCenter = getProceduresPerCenter();
-
-		for(String center: centers){
-			List<StrainBean> strains = getMutantStrainsForCenter(center);
-
-			for(StrainBean strain: strains){
-				List<String> procedures = getDoneProcedureIdsPerStrainAndCenter(center, strain.getColonyId());
-				List<String> row = new ArrayList<>();
-				row.add(strain.getColonyId());
-				row.add(strain.getGeneSymbol());
-				row.add(strain.getMgiAccession());
-				row.add(strain.getAllele());
-
-				row.add(strain.getZygosity());
-				row.add(strain.getLifeStage());
-				row.add(center);
-				Float percentageDone = (float) ((procedures.size() * 100) / (float)possibleProceduresPerCenter.get(center).size());
-				row.add(percentageDone.toString());
-				row.add("" + procedures.size()); // #procedures done
-				row.add(procedures.toString()); // procedures done
-				Set<String> mpsTestedFor = getMpsForProcedureSet(procedures, mpsPerProcedure);
-				row.add(mpsTestedFor.toString()); //mpsTestedFor
-				row.add("" + (possibleProceduresPerCenter.get(center).size() - procedures.size()));	// #missing procedures
-				List<String> missing = new ArrayList<>(possibleProceduresPerCenter.get(center));
-				missing.removeAll(procedures); // missing procedures
-				row.add(missing.toString());
-				Set<String> mpsNotTestedFor = getMpsForProcedureSet(missing, mpsPerProcedure);
-				mpsNotTestedFor.removeAll(mpsTestedFor);
-				row.add(mpsNotTestedFor.toString());
-				results.add(row.toArray(temp));
-			}
-		}
-		return results;
-	}
-
-	private Set<String> getMpsForProcedureSet(List<String> procedures, Map<String, Set<String>> mpsPerProcedure){
+	public Set<String> getMpsForProcedureSet(List<String> procedures, Map<String, Set<String>> mpsPerProcedure){
 
 		HashSet<String> res = new HashSet<>();
 		for (String procedure: procedures){
@@ -334,75 +260,5 @@ public class PhenotypeCenterService {
 			}
 		}
 		return res;
-	}
-
-}
-
-class StrainBean{
-
-	String mgiAccession;
-	String allele;
-	String geneSymbol;
-	String colonyId;
-	String zygosity;
-	String lifeStage;
-
-
-	public String getMgiAccession() {
-
-		return mgiAccession;
-	}
-
-	public void setMgiAccession(String mgiAccession) {
-
-		this.mgiAccession = mgiAccession;
-	}
-
-	public String getAllele() {
-
-		return allele;
-	}
-
-	public void setAllele(String allele) {
-
-		this.allele = allele;
-	}
-
-	public String getGeneSymbol() {
-
-		return geneSymbol;
-	}
-
-	public void setGeneSymbol(String geneSymbol) {
-
-		this.geneSymbol = geneSymbol;
-	}
-
-
-	public String getColonyId() {
-
-		return colonyId;
-	}
-
-
-	public void setColonyId(String colonyId) {
-
-		this.colonyId = colonyId;
-	}
-
-	public String getZygosity() {
-		return zygosity;
-	}
-
-	public void setZygosity(String zygosity) {
-		this.zygosity = zygosity;
-	}
-
-	public String getLifeStage() {
-		return lifeStage;
-	}
-
-	public void setLifeStage(String lifeStage) {
-		this.lifeStage = lifeStage;
 	}
 }
