+++ reports/src/main/java/org/mousephenotype/cda/reports/ProcedureCompletenessReport.java
@@ -19,10 +19,9 @@ package org.mousephenotype.cda.reports;
 import org.apache.commons.lang3.ClassUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.mousephenotype.cda.reports.support.ReportException;
-import org.mousephenotype.cda.solr.service.PhenotypeCenterService;
+import org.mousephenotype.cda.solr.service.PhenotypeCenterProcedureCompletenessService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.beans.Introspector;
@@ -37,13 +36,15 @@ import java.util.List;
 @Component
 public class ProcedureCompletenessReport extends AbstractReport {
 
-    protected Logger logger = LoggerFactory.getLogger(this.getClass());
+    protected Logger                                      logger = LoggerFactory.getLogger(this.getClass());
+    private   PhenotypeCenterProcedureCompletenessService phenotypeCenterProcedureCompletenessService;
 
-    @Autowired
-    PhenotypeCenterService phenotypeCenterService;
 
-    public ProcedureCompletenessReport() {
+    public ProcedureCompletenessReport(
+            PhenotypeCenterProcedureCompletenessService phenotypeCenterProcedureCompletenessService
+    ) {
         super();
+        this.phenotypeCenterProcedureCompletenessService = phenotypeCenterProcedureCompletenessService;
     }
 
     @Override
@@ -64,7 +65,7 @@ public class ProcedureCompletenessReport extends AbstractReport {
 
         List<String[]> result;
         try {
-            result = phenotypeCenterService.getCentersProgressByStrainCsv();
+            result = phenotypeCenterProcedureCompletenessService.getCentersProgressByStrainCsv();
         } catch (Exception e) {
             throw new ReportException("Exception creating " + this.getClass().getCanonicalName() + ". Reason: " + e.getLocalizedMessage());
         }
