+++ reports/src/main/java/org/mousephenotype/cda/reports/support/SolrServerConfig.java
@@ -3,6 +3,7 @@ package org.mousephenotype.cda.reports.support;
 import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.mousephenotype.cda.solr.service.ImpressService;
+import org.mousephenotype.cda.solr.service.PhenotypeCenterProcedureCompletenessService;
 import org.mousephenotype.cda.solr.service.PhenotypeCenterService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -148,9 +149,18 @@ public class SolrServerConfig {
 		return new HttpSolrClient(solrBaseUrl + "/statistical-result");
 	}
 
-	@Bean(name = "phenotypeCenterService")
+
+	// Service Beans
+
+
+	@Bean
 	PhenotypeCenterService phenotypeCenterService() {
-		return new PhenotypeCenterService(solrBaseUrl + "/experiment", impressService());
+		return new PhenotypeCenterService(getExperimentCore());
+	}
+
+	@Bean
+	PhenotypeCenterProcedureCompletenessService phenotypeCenterProcedureCompletenessService() {
+		return new PhenotypeCenterProcedureCompletenessService(phenotypeCenterService(), impressService(), getStatisticalResultCore());
 	}
 
 }
