+++ web/src/main/java/uk/ac/ebi/phenotype/SolrServerConfig.java
@@ -3,6 +3,7 @@ package uk.ac.ebi.phenotype;
 import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.mousephenotype.cda.solr.service.ImpressService;
+import org.mousephenotype.cda.solr.service.PhenotypeCenterProcedureCompletenessService;
 import org.mousephenotype.cda.solr.service.PhenotypeCenterService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -159,22 +160,23 @@ public class SolrServerConfig {
         return new HttpSolrClient(solrBaseUrl + "/preqc");
     }
 
-
-//	@Bean(name = "preqcCore") //DELETEME
-//	HttpSolrClient getPreqcCore() {
-//		return new HttpSolrClient(solrBaseUrl + "/genotype-phenotype");
-//	}
-
     //StatisticalResult
     @Bean(name = "statisticalResultCore")
     HttpSolrClient getStatisticalResultCore() {
         return new HttpSolrClient(solrBaseUrl + "/statistical-result");
     }
 
-    @Bean(name = "phenotypeCenterService")
+
+    // Service Beans
+
+
+    @Bean
     PhenotypeCenterService phenotypeCenterService() {
-        return new PhenotypeCenterService(solrBaseUrl + "/experiment", impressService);
+        return new PhenotypeCenterService(getExperimentCore());
     }
 
-
+    @Bean
+    PhenotypeCenterProcedureCompletenessService phenotypeCenterProcedureCompletenessService() {
+        return new PhenotypeCenterProcedureCompletenessService(phenotypeCenterService(), impressService, getStatisticalResultCore());
+    }
 }
